feat(iam): improve disable credentials checks (#2909)
