+++ prowler/compliance/aws/mitre_attack_aws.json
@@ -167,7 +167,8 @@
         "iam_customer_unattached_policy_no_administrative_privileges",
         "iam_inline_policy_no_administrative_privileges",
         "iam_no_expired_server_certificates_stored",
-        "iam_disable_30_days_credentials",
+        "iam_user_accesskey_unused",
+        "iam_user_console_access_unused",
         "iam_no_root_access_key",
         "iam_no_custom_policy_permissive_role_assumption",
         "iam_password_policy_expires_passwords_within_90_days_or_less",
@@ -673,7 +674,8 @@
         "iam_policy_allows_privilege_escalation",
         "iam_policy_no_full_access_to_cloudtrail",
         "iam_policy_no_full_access_to_kms",
-        "iam_disable_30_days_credentials",
+        "iam_user_accesskey_unused",
+        "iam_user_console_access_unused",
         "iam_no_root_access_key",
         "iam_root_hardware_mfa_enabled",
         "iam_root_mfa_enabled",
@@ -720,7 +722,8 @@
       "Checks": [
         "guardduty_is_enabled",
         "config_recorder_all_regions_enabled",
-        "iam_disable_30_days_credentials",
+        "iam_user_accesskey_unused",
+        "iam_user_console_access_unused",
         "iam_password_policy_expires_passwords_within_90_days_or_less",
         "iam_password_policy_lowercase",
         "iam_password_policy_minimum_length_14",
