+++ tests/lib/check/fixtures/groupsA.json
@@ -9,8 +9,8 @@
     },
     "iam": {
       "checks": [
-        "iam_disable_30_days_credentials",
-        "iam_disable_90_days_credentials"
+        "iam_user_accesskey_unused",
+        "iam_user_console_access_unused"
       ],
       "description": "Identity and Access Management"
     }
